requirements.txt
torch>=2.1.0
transformers>=4.41.0
tiktoken>=0.7.0
Pillow>=10.0.0
requests>=2.31.0
tqdm>=4.66.0
accelerate>=0.32.0
safetensors>=0.4.3
transformers>=4.42.0
sentence-transformers
torchvision
huggingface_hub